+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 ## [Unreleased]
 
+## [0.2.2] - 2025-08-08
+### Added
+- Added setting for notification timeout.
+
+### Changed
+- Changed behavior of the excluded folder, now these folders are interpreted as paths relative to the root of the project.
+
 ## [0.2.1] - 2025-08-08
 ### Added
-- Added expiration time (3s) for the notification, after that time notification will auto-hide.
+- Added expiration time (3s) for the notification, after that time the notification will auto-hide.
 
 ## [0.2.0] - 2025-08-05
 ### Added